Make bitmap.c all inline
The very short functions in bitmap.c are called extremely frequently. Having them  inlined would likely allow further optimizations.

Now this change would not be needed if the program was built with -flto, but (1) that is not what the Makefile says (2) doing this change is such a very-low-hanging fruit that bringing in the big gun of LTO (with the compile time increases) is not exactly justified.